+++ Autodesk.Forge.Oss/OssClientFileExtension.cs
@@ -23,6 +23,57 @@ namespace Autodesk.Forge.Oss
         /// <returns>The ObjectDetails of the uploaded file.</returns>
         public static async Task<ObjectDetails> UploadFileAsync(this OssClient oss, string bucketKey, string objectName, string localFullName)
         {
+            // 2MB is minimal, clamp to it
+            int chunkMbSize = 12; // 12/2 => 6.0 > 5MB is minimal
+            long chunkSize = chunkMbSize * 1024 * 1024;
+            int chunkNumber = 1;
+
+            using var fileReadStream = File.OpenRead(localFullName);
+
+            // determine if we need to upload in chunks or in one piece
+            long sizeToUpload = fileReadStream.Length;
+
+            // Auto Chunk Size
+            if (sizeToUpload > chunkSize)
+            {
+                chunkNumber = (int) Math.Ceiling((double)sizeToUpload / (double)chunkSize);
+                chunkSize = (long)Math.Ceiling(((double)sizeToUpload / chunkNumber));
+            }
+
+            var signeds3uploadResponse = await oss.GetS3UploadURLAsync(bucketKey, objectName, chunkNumber);
+
+            // use chunks for all files greater than chunk size
+            if (sizeToUpload > chunkSize)
+            {
+                string sessionId = Guid.NewGuid().ToString();
+                long begin = 0;
+                int urlIndex = 0;
+                byte[] buffer = new byte[chunkSize];
+
+                while (begin < sizeToUpload - 1)
+                {
+                    int memoryStreamSize = sizeToUpload - begin < chunkSize ? (int)(sizeToUpload - begin) : (int)chunkSize;
+                    var bytesRead = await fileReadStream.ReadAsync(buffer, 0, memoryStreamSize);
+                    using var chunkStream = new MemoryStream(buffer, 0, memoryStreamSize);
+
+                    using HttpClient httpClient = new HttpClient();
+                    using StreamContent streamContent = new StreamContent(chunkStream);
+                    HttpResponseMessage response = await httpClient.PutAsync(signeds3uploadResponse.urls[urlIndex++], streamContent);
+
+                    begin += bytesRead;
+                }
+            }
+            else
+            {
+                var url = signeds3uploadResponse.urls.FirstOrDefault();
+                using HttpClient httpClient = new HttpClient();
+                using StreamContent streamContent = new StreamContent(fileReadStream);
+                HttpResponseMessage response = await httpClient.PutAsync(url, streamContent);
+            }
+            return await oss.CompleteS3UploadAsync(bucketKey, objectName, signeds3uploadResponse.uploadKey, (int)sizeToUpload);
+        }
+        internal static async Task<ObjectDetails> UploadFileAsync2(this OssClient oss, string bucketKey, string objectName, string localFullName)
+        {
             var signeds3uploadResponse = await oss.GetS3UploadURLAsync(bucketKey, objectName);
             var url = signeds3uploadResponse.urls.FirstOrDefault();